+++ src/main/java/com/milaboratory/mir/mappers/stm/ExplicitAlignmentScoring.java
@@ -4,6 +4,5 @@ import com.milaboratory.core.mutations.Mutations;
 import com.milaboratory.core.sequence.Sequence;
 
 public interface ExplicitAlignmentScoring<S extends Sequence<S>> {
-    double computeScore(S query,
-                        Mutations<S> mutations);
+    float computeScore(S query, Mutations<S> mutations);
 }
